requirements.txt
asgiref==3.8.1
attrs==25.1.0
azure-cognitiveservices-speech==1.42.0
certifi==2025.1.31
colorama==0.4.6
Django==5.1.7
django-authtools==2.0.1
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nltk==3.9.1
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.10.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rpds-py==0.22.3
sentry-sdk==2.22.0
sqlparse==0.5.3
typing_extensions==4.12.2
tqdm==4.67.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
